[6.x] Move Control Panel controllers (#372)
* Move Control Panel controllers

* Fix styling

* Move form requests too

---------
